+++ contracts/IColony.sol
@@ -13,7 +13,7 @@ contract IColony {
   function setToken(address _token) public;
   function getToken() public view returns (address);
   function initialiseColony(address _network) public;
-  function mintTokens(uint128 _wad) public;
+  function mintTokens(uint256 _wad) public;
   function addSkill(uint256 _parentSkillId) public;
 
   // ColonyTask
